requirements.txt
pandas
psycopg2-binary
openpyxl
matplotlib
tqdm